vscode/pyproject.toml
[tool.poetry]

name = "demo"

version = "0.1.0"

description = "An example project demonstrating my preferred setup for a Python project."

license = ""

authors = ["John Doe <johndoe@example.com>"]

maintainers = [
    "George Washington <georgewashington@example.com>",
    "Jane Doe <janedoe@example.com>",
]

readme = "README.md"

homepage = "https://github.com/sandmanscanga/Misc/blob/master/README.md"

repository = "https://github.com/sandmanscanga/Misc"

keywords = ["app", "example", "project", "python"]

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Natural Language :: English",
    "Topic :: Software Development",
    "Typing :: Typed",
]

packages = [{ include = "demo", from = "." }]

include = ["tests/", "pyproject.toml", "README.md"]

exclude = [
    "**/__pycache__/",
    "**/*.pyc",
    "build/",
    "dist/",
    "docs/",
    "venv/",
    "*.code-workspace",
    "*.tar.gz",
    "build.sh",
    "Jenkinsfile",
    "Operationsfile",
    "poetry.lock",
    "ver.config",
    ".cspell/",
    ".env",
    ".git/",
    ".gitattributes",
    ".gitignore",
    ".idea/",
    ".mypy_cache/",
    ".pre-commit-config.yaml",
    ".pylintrc",
    ".python-version",
    ".pytest_cache/",
    ".venv/",
    ".vscode/",
]

[tool.poetry.scripts]
run_unit_tests = "demo.scripts.run_unit_tests:main"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.3"

[tool.poetry.group.dev.dependencies]
bandit = "^1.7.8"
black = "^24.4.0"
blacken-docs = "^1.16.0"
bumpversion = "^0.6.0"
colorama = "^0.4.6"
colorlog = "^6.8.2"
ipdb = "^0.13.13"
ipython = "^8.23.0"
isort = "^5.13.2"
mypy = "^1.9.0"
notebook = "^7.1.3"
pre-commit = "^3.7.0"
psutil = "^5.9.8"
pylint = "^3.1.0"
python-dotenv = "^1.0.1"
rich = "^13.7.1"


[tool.poetry.group.testing.dependencies]
coverage = "^7.4.4"
pygments-pytest = "^2.4.0"
pytest = "^8.1.1"
pytest-benchmark = "^4.0.0"
pytest-cov = "^5.0.0"
pytest-emoji = "^0.2.0"
pytest-logdog = "^0.1.0"
pytest-menu = "^0.0.4"
pytest-mock = "^3.14.0"
pytest-pretty = "^1.2.0"
pytest-sorter = "^0.3.1"
pytest-threadleak = "^0.5.0"


[tool.poetry.group.docs.dependencies]
click-man = "^0.4.1"
doc-utils = "^0.0.2"
mkdocs = "^1.6.0"
mkdocstrings = "^0.24.3"
mkdocstrings-python = "^1.10.0"
pdoc3 = "^0.10.0"
pydocstyle = "^6.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
